spank: refuse to load the same plugin more than once
Avoid loading the same plugin more than once in plugstack.c.
Most likely this will be a configuration error, so we should
catch it early. If the same .so appears in the plugin stack
more than once, it is likely to cause very strange errors,
since dlopen() will only map the library a single time.