spank: remove spank_set_remote_options_env
The spank_set_remote_options_env() function is not used anywhere except
internal to plugstack.c, so remove it from plugstack.h. Then redefine
it to take a spank_stack argument so that it doesn't refer to the
global_spank_stack. Finally rename to spank_stack_set_remote_options_env()
to clarify the intent.