spank: avoid loading plugins with no callbacks for current context
In slurmd and job prolog/epilog contexts, avoid loading plugins that
have no callbacks in the context in which they are loaded. That is
for slurmd, if there are no slurm_spank_slurmd_init or
slurm_spank_slurmd_exit callbacks, there is no reason to keep the
current plugin loaded.