Made changes to allow --export-file option to get missing environment variables.
Our original premise that environment data from the file designated by
the --export-file option should be the only elements of the user environemnt
was wrong. The code needed, as it did previously, continue to merge the
user's environment that is established by the sbatch command at initiation
by the user. This required adding a parameter to env_array_merge so that
it could 'add' environment variables without overwriting an variable that
already existed.