Unset user's environment at start of execution if --export-file option..
Because the --export-file option is indicating the user's environment
is comming from a file, the environemnt, as set in "extern char **environ,
is now unset to insure that those variables do not overwrite the ones
imported from the file. This is consistent with the how it works when
spawned by Moab.