slurmd: Refactor code to run prolog/epilog
We now want to return error on failure of either spank prolog/epilog
or regular prolog/epilog scripts, so add a common function _run_job_script
to handle return of shared error code.

For now, we continue to run the normal prolog or epilog even if the spank
prolog/epilog fail. In the future, a failure the spank prolog/epilog may
short-circuit the run of the normal scripts.