slurmd: remove log call between fork and exec in run_spank_job_script
Remove potentially unsafe call to info() between fork and exec in
run_spank_job_script(). This call has the potential to cause a deadlock
if another thread has a mutex locked at the time of fork(), and there is
no reason the log call had to be made from the child process.