Patch to fix perl api module node.c
The perl api call to load_node will take an
argument, "SHOW_ALL", which will result in
returning all node names, even those in a
hidden partition. If called without an agrument
it should respond like the c api, and return all
a NULL node name for those nodes in the hidden
oartition.

However, the node.c module was returning a error
when encountering a NULL node name, this fix
checks for NULL and skips the record.

SchedMD is making the same mod.