Enable build of slurm debuginfo packages
The build of debuginfo packages was disabled presumably to ensure
that debug symbols are not stripped from srun, so that debuggers
like TotalView can attach. However, as long as we do not completely
strip the binaries, the MPIR_* should be accessible, so we should
not have to completely disable debuginfos.